# perf: improve typescript type checking performance