@andreialecu 2021-07-30
I have a related question. 

Is it actually ok to allow strings for object ids in query filters? In my experience trying to query something as a string when it's stored as an `ObjectId` doesn't return any results. I always had to compare it with `someId: new ObjectId(somestring)`.

Are the types actually correct in trying to allow strings for object ids for filters here?