@perenSHERE 2021-07-30
> **This PR might be able to fix some issues but not issues with the systems who have half decent specification**

**my system specifications are:**
- i5 2500k 
- 8 GB ram
- gt 1030

**before applying changes in the PR #10515**
```
Files:                          87
Lines of Library:            28119
Lines of Definitions:        25937
Lines of TypeScript:            19
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Nodes of Library:           119708
Nodes of Definitions:       108811
Nodes of TypeScript:            65
Nodes of JavaScript:             0
Nodes of JSON:                   0
Nodes of Other:                  0
Identifiers:                 81545
Symbols:                    128399
Types:                       43073
Instantiations:             291774
Memory used:               149894K
Assignability cache size:    13123
Identity cache size:             3
Subtype cache size:              0
Strict subtype cache size:       0
I/O Read time:               0.02s
Parse time:                  0.64s
ResolveModule time:          0.07s
ResolveTypeReference time:   0.01s
Program time:                0.76s
Bind time:                   0.36s
Check time:                  1.87s
transformTime time:          0.01s
commentTime time:            0.00s
I/O Write time:              0.00s
printTime time:              0.02s
Emit time:                   0.02s
Total time:                  3.02s
```
**after applying changes in the PR #10515**
```
Files:                          87
Lines of Library:            28119
Lines of Definitions:        25937
Lines of TypeScript:            19
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Nodes of Library:           119708
Nodes of Definitions:       108811
Nodes of TypeScript:            65
Nodes of JavaScript:             0
Nodes of JSON:                   0
Nodes of Other:                  0
Identifiers:                 81545
Symbols:                     98716
Types:                       23615
Instantiations:             153096
Memory used:               125551K
Assignability cache size:     7426
Identity cache size:             1
Subtype cache size:              0
Strict subtype cache size:       0
I/O Read time:               0.02s
Parse time:                  0.63s
ResolveModule time:          0.06s
ResolveTypeReference time:   0.01s
Program time:                0.74s
Bind time:                   0.36s
Check time:                  1.40s
transformTime time:          0.01s
commentTime time:            0.00s
I/O Write time:              0.00s
printTime time:              0.02s
Emit time:                   0.02s
Total time:                  2.51s
```
> **This PR gives a bit more boost in the performance but its still not enough**