@andreialecu 2021-07-30
@perenSHERE the check time itself is not necessarily conclusive to VS code performance. But it's a decent indicator to check whether changes are for the better or worse.

Did you try this in VS code?