@perenSHERE 2021-07-30
> @perenSHERE the check time itself is not necessarily conclusive to VS code performance. But it's a decent indicator to check whether changes are for the better or worse.
> 
> Did you try this in VS code?

Yes I tried this in VS code 
```
Version: 1.58.2 (user setup)
Commit: c3f126316369cd610563c75b1b1725e0679adfb3
Date: 2021-07-14T22:10:15.214Z
Electron: 12.0.13
Chrome: 89.0.4389.128
Node.js: 14.16.0
V8: 8.9.255.25-electron.0
OS: Windows_NT x64 10.0.19043
```