@andreialecu 2021-07-30
It appears that changing from:

```
  class Schema<DocType = Document, M extends Model<DocType, any, any> = Model<any, any, any>, SchemaDefinitionType = undefined, TInstanceMethods = ExtractMethods<M>> extends events.EventEmitter {
```

to (notice not supplying a default `Model`):

```
  class Schema<DocType = Document, M extends Model<DocType, any, any> = any, SchemaDefinitionType = undefined, TInstanceMethods = ExtractMethods<M>> extends events.EventEmitter {
```

results in this trace:

```
Nodes of Other:                 0
Identifiers:                82981
Symbols:                    53712
Types:                       3337
Instantiations:              5223
Memory used:               98405K
Assignability cache size:    1611
Identity cache size:           11
Subtype cache size:             0
Strict subtype cache size:      0
I/O Read time:              0.01s
Parse time:                 0.38s
ResolveModule time:         0.02s
ResolveTypeReference time:  0.00s
Program time:               0.42s
Bind time:                  0.19s
Check time:                 0.08s
transformTime time:         0.00s
commentTime time:           0.00s
I/O Write time:             0.00s
printTime time:             0.01s
Emit time:                  0.01s
Total time:                 0.71s
```

I'm not sure why this happens, pretty bizarre. Still investigating.