@andreialecu 2021-07-30
If I leave all the code in `Schema` but remove all the code from the `Model` interface, the check time drops to 0.05. If I only leave one of the `update*` lines like the one below, it shoots back up to 1+ seconds:

```ts
updateOne(filter?: FilterQuery<T>, update?: UpdateQuery<T> | UpdateWithAggregationPipeline, options?: QueryOptions | null, callback?: (err: CallbackError, res: any) => void): QueryWithHelpers<UpdateWriteOpResult, EnforceDocument<T, TMethods>, TQueryHelpers, T>;
```

If I remove its return type, the check time goes down to `0.2s`.

---

Alternatively, if everything is kept but the code of `Query` is commented out, the check time goes down to `0.34s`.

So the problem isn't in the `Schema` type, but somewhere deeper down.