@andreialecu 2021-07-30
Update: Changing the definition of `Schema` to this:

```ts
class Schema<DocType = Document, M = Model<any, any, any>, SchemaDefinitionType = undefined, TInstanceMethods = any> extends events.EventEmitter {
```

Results in this trace.

```
Instantiations:              1029
Memory used:               97021K
Assignability cache size:     161
Identity cache size:            0
Subtype cache size:             0
Strict subtype cache size:      0
I/O Read time:              0.01s
Parse time:                 0.38s
ResolveModule time:         0.02s
ResolveTypeReference time:  0.00s
Program time:               0.43s
Bind time:                  0.20s
Check time:                 0.04s
```

It appears that one potential problem is with the `extends Model<DocType, any, any>` check. `ExtractMethods<T>` also has one.