@andreialecu 2021-07-30
I have opened https://github.com/microsoft/TypeScript/issues/45249, perhaps someone can offer some guidance.