@perenSHERE 2021-07-30
> Update: Changing the definition of `Schema` to this:
> 
> ```ts
> class Schema<DocType = Document, M = Model<any, any, any>, SchemaDefinitionType = undefined, TInstanceMethods = any> extends events.EventEmitter {}
> It appears that one potential problem is with the `extends Model<DocType, any, any>` check. `ExtractMethods<T>` also has one.

Definitely commit this to the branch dropped the Total Time Decently for me

**I am not that experienced with type declaration but after reading some code Model Interface can be optimized decently**
``` interface Model<T, TQueryHelpers = {}, TMethods = {}> extends NodeJS.E1ventEmitter, AcceptsDiscriminator {}```